spark-backend/requirements.txt
fastapi[all]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
python-multipart
cloudinary
pytest
pytest-cov
httpx